FFCProcessor Class

The new workhorse of the client - this class is meant to bring together the scattershot functions in other parts of the package so that data can be integrated into a single class with a single set of tasks. Other functions are likely to be supported for a while (and this may even rely on them), but long run, much of the code in this file might move into this class, with the shortcut functions creating this class behind the scenes and returning an instance of this object.

Details

More details to come, and more examples. For now, still use the general functions evaluate_alteration and evaluate_gage_alteration

Get Gage ID for FFCProcessor

We may not always have a gage ID, but may want to just get one if it exists - this function gets a gage ID if we're using a gage or returns NA otherwise

Provides alteration scores

Checks the results against the predictions and returns the appropriate alteration score

Methods

Public methods

FFCProcessor$set_up()

FFCProcessor$run()

FFCProcessor$rename_inconsistent_metrics()

FFCProcessor$get_gage_id()

FFCProcessor$step_one_functional_flow_results()

FFCProcessor$step_two_explore_ecological_flow_criteria()

FFCProcessor$step_three_assess_alteration()

FFCProcessor$get_ffc_results()

FFCProcessor$evaluate_alteration()

FFCProcessor$clone()

Method set_up()

Usage

FFCProcessor$set_up(gage_id, timeseries, comid, token)

Method run()

Usage

FFCProcessor$run()

Method rename_inconsistent_metrics()

Usage

FFCProcessor$rename_inconsistent_metrics()

Method get_gage_id()

Usage

FFCProcessor$get_gage_id()

Method step_one_functional_flow_results()

Usage

FFCProcessor$step_one_functional_flow_results(
  gage_id,
  timeseries,
  comid,
  token,
  output_folder
)

Method step_two_explore_ecological_flow_criteria()

Usage

FFCProcessor$step_two_explore_ecological_flow_criteria()

Method step_three_assess_alteration()

Usage

FFCProcessor$step_three_assess_alteration()

Method get_ffc_results()

Usage

FFCProcessor$get_ffc_results()

Method evaluate_alteration()

Usage

FFCProcessor$evaluate_alteration()

Method clone()

The objects of this class are cloneable with this method.

Usage

FFCProcessor$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.