USGS Gage Retrieval Tools

This class retrieves data for a USGS gage.

Details

Methods

Public methods

USGSGage$validate()

USGSGage$get_data()

USGSGage$get_comid()

USGSGage$get_predicted_metrics()

USGSGage$clone()

Method validate()

Usage

USGSGage$validate(latlong)

Details

Validates that gage is ready to run requests

Internal method. Checks parameters to make sure they're ready for other methods on the object.

Method get_data()

Usage

USGSGage$get_data()

Method get_comid()

Usage

USGSGage$get_comid(online = TRUE)

Details

Looks up the COMID for this gage

This method looks up the COMID for the gage and sets the comid attribute. It does not return the COMID. It returns this object for chaining. The gage's id, latitude, and longitude attributes must be set before running this. latitude and longitude can be set manually, or by running get_data(). Can be error prone near stream junctions. If you have the means to get a reliable COMID for a gage, do so - in this method, we look up the stream segment by long/lat using nhdPlusTools.

Method get_predicted_metrics()

Usage

USGSGage$get_predicted_metrics(force_comid_lookup)

Method clone()

The objects of this class are cloneable with this method.

Usage

USGSGage$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples